+++ activities/ActivityTemplate/VueJS/js/components/SugarPresence.js
@@ -85,9 +85,34 @@ Vue.component('sugar-presence', {
 			this.$emit('user-changed', msg);
 
 			console.log("User " + msg.user.name + " " + (msg.move == 1 ? "joined" : "left"));
+			let vm = this;
+			// Show popup
 			if (this.$root.$refs.SugarPopup) {
-				this.$root.$refs.SugarPopup.log("User " + msg.user.name + " " + (msg.move == 1 ? "joined" : "left"));
+				let stringToDisplay = '';
+
+				// Get localized string
+				if(this.$root.$refs.SugarL10n) {
+					let userStringToGet = "User" + (msg.move == 1 ? "Joined" : "Left");
+					stringToDisplay = this.$root.$refs.SugarL10n.get(userStringToGet, { name: msg.user.name });
+				} else {
+					stringToDisplay = "User " + msg.user.name + " " + (msg.move == 1 ? "joined" : "left");
+				}
+
+				// Show icon
+				if(this.$root.$refs.SugarIcon) {
+					this.$root.$refs.SugarIcon.generateIconWithColors('../icons/owner-icon.svg', msg.user.colorvalue)
+						.then(src => {
+							let html = `<div style="display: flex; align-items:center;">`;
+							let img = "<img style='height:40px; margin:10px; vertical-align: middle;' src='" + src + "'>";
+							html += `${img}`;
+							html += `<span>${stringToDisplay}</span>`
+							html += `</div>`;
+							vm.$root.$refs.SugarPopup.log(html);
+						});
+				} else {
+					this.$root.$refs.SugarPopup.log(stringToDisplay);
+				}
 			}
-		},
+		}
 	}
 });