+++ activities/MindMath.activity/js/activity.js
@@ -938,6 +938,12 @@ var app = new Vue({
 
     onJournalLoadError: function(error) {
       console.log("Error loading from journal");
+		},
+		
+    onActivityShared: function(event, paletteObject) {
+      this.onMultiplayerGameStarted();
+      // Usual behaviour call
+      this.SugarPresence.onShared(event, paletteObject);
     },
 
     onNetworkDataReceived: function(msg) {